feat: 添加 gRPC-Gateway 集成和 API 重构
将 QA 服务从 Gin HTTP 迁移到 gRPC 架构，并引入 gRPC-Gateway 实现 RESTful API 支持。

主要变更:
- 为 proto 文件添加 HTTP 注解和 field_mask 支持
- 实现 QA 服务的完整 gRPC handler，包括问题、回答和评论的 CRUD 操作
- 添加点赞/取消点赞功能的 gRPC 接口
- 配置 gRPC-Gateway 作为统一网关，代理 User 和 QA 服务
- 重构身份验证中间件，支持 gRPC 拦截器和公共方法白名单
- 改进用户身份信息结构，添加 Claims 字段以支持完整的 JWT 访问
- 更新 protoc 生成命令，包含 googleapis 依赖
- 从请求参数中移除 user_id，改为从认证上下文获取
- 将分页参数类型优化为 int64/int32
- 添加 Logout 接口到用户服务

技术优化:
- 统一使用 gRPC 作为服务间通信协议
- 通过 gRPC-Gateway 自动生成 RESTful API
- 简化前端 API 路径（从 /users 改为 /auth）
- 增强认证中间件的灵活性和可配置性

+++ cmd/qa-service/internal/handler/gin_handler.go
@@ -121,7 +121,7 @@ func getIDFromParam(c *gin.Context, paramName string) (int64, bool) {
 	return id, true
 }
 
-func getPagination(c *gin.Context) (int, int) {
+func getPagination(c *gin.Context) (int64, int32) {
 	page, _ := strconv.Atoi(c.DefaultQuery("page", "1"))
 	pageSize, _ := strconv.Atoi(c.DefaultQuery("pageSize", "10"))
 	if page < 1 {
@@ -130,7 +130,7 @@ func getPagination(c *gin.Context) (int, int) {
 	if pageSize < 1 {
 		pageSize = 10
 	}
-	return page, pageSize
+	return int64(page), int32(pageSize)
 }
 
 // --- Question Handlers ---
